Commit 30
- Melhorias na persistência das entidades na requisição dos controller;
- Remoção de código que não será mais utilizado;
- Implementação do vínculo do carro com a vaga;
- Implementação de uma classe abrstrata de DTOs;
- Adição de SQL para criar 200 vagas no banco de dados na inicialização do projeto.

+++ src/main/java/marhlonkorb/github/io/gerenciadorestacionamento/models/repositories/VagaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface VagaRepository extends JpaRepository<Vaga, Integer> {
+public interface VagaRepository extends JpaRepository<Vaga, Long> {
 }
